### 2021-01-24 #2984 @whedon
[PRE REVIEW]: Bayesian X-ray Analysis (BXA) v4.0
**Submitting author:** @JohannesBuchner (<a href="http://orcid.org/0000-0003-0426-6634">Johannes Buchner</a>)
**Repository:** <a href="https://github.com/JohannesBuchner/BXA" target ="_blank">https://github.com/JohannesBuchner/BXA</a>
**Version:** v4.0.2
