### 2021-01-29 c19c584
Corrected comment on prior from file
Corrected the misleading comment on how to format the file to create a prior for a parameter.
