### 2021-02-15 #2984 @JohannesBuchner
Here are some highly qualified people who know X-rays and statistics and software packaging, but I don't know if they can donate time to be reviewers: @dhuppenkothen @cescalara @giacomov @grburgess and perhaps [stingray contributors](https://github.com/StingraySoftware/stingray/blob/main/CREDITS.rst).
