### 2021-03-07 JohannesBuchner/BXA#19 @cescalara
JOSS review: documentation
Hi, a few comments/questions on the documentation as part of the review:

- [x] Installation: Add the full list of dependencies to the install steps. I also needed `cython`, `tqdm`, `h5py` and `ultranest` as a minimal list to get things running.
- [x] For both the sherpa/Xspec cases, it would be nice to see a full standalone demo including some example data. I know this is done in the scripts in `examples/` , but would be helpful if at least one case was treated notebook-style on the documentation pages.
- [x] Similarly for the `xagnfitter.py` script - you have a case with some example data in the Travis CI, it could be nice to also have this in the docs as a walkthrough for users.
- [x] The default ultranest output is cool, but could be distracting/annoying when running in certain settings. Could you show in the docs an easy interface in BXA to switch it off?
- [x] In the "convenience features" documentation
  - [x] Seems the example call of `gal.py` is a typo as refers to another script
  - [x] Would be nice to have some quick code showing exactly how to import chains in Xspec or run a script in parallel
