### 2021-03-07 JohannesBuchner/BXA#20 @cescalara
JOSS review: examples/tests
Hi, some more comments on the example code:

- [x] It seems that some of the sherpa examples rely on `npyinterp` and python 2? Could this be updated to python 3? Please correct me if I am wrong.
- [x] The example code in `docker/` also seems a bit outdated (e.g. doesn't use `ultranest` and python 3). OK if you want to keep this for reference, but then an up-to-date docker example should be added too.
- [x] What is the difference between `xagnfitter.py` detailed in the docs and `fitagn.py` in the docker examples?
- [x] When running `gal.py` I get a connection timeout to swift.ac.uk (looks likely to be a temporary problem on their side, so just to notify you)
- [x] Can you fix the latest Travis CI build? Seems that the examples run fine and it should work
