### 2021-03-07 #3045 @cescalara
Hi @JohannesBuchner, Thanks for all your work on BXA! Overall it looks like things are working and well-documented, but I do have a few comments/questions. I have opened a couple of issues dealing with code-related things:

* Documentation - [#19](https://github.com/JohannesBuchner/BXA/issues/19)
* Examples/tests - [#20](https://github.com/JohannesBuchner/BXA/issues/20)

And also a few comments on the paper:

* What is the relationship to [pyblocxs](https://github.com/brefsdal/pyblocxs)? I realise that this package is no longer active, and little of the original code is left. however, if this code contributed to the ideas and development of BXA, it seems proper to note this and acknowledge the author in the spirit of good open-source etiquette.
* For me, it would be useful to have a high-level description of Xspec/sherpa and their differences in the paper. It would also make sense to mention [3ML](https://github.com/threeML/threeML) as an alternative tool for X-ray analysis and the differences there.
* Would be nice to have a short overview of the methods/functionality in the paper in addition to pointing to your 2014 paper for details.

I really hope my comments are helpful and I'm happy to discuss on any of the points.

Cheers!
