+++ .travis.yml
@@ -143,7 +143,7 @@ install:
 
   - |
     if [[ "$TOOL" != "sherpa-conda" ]]; then
-      sudo pip3 install requests corner astropy h5py cython coveralls coverage scipy tqdm
+      sudo pip3 install requests corner astropy h5py cython coveralls coverage scipy tqdm;
       sudo pip3 install ultranest
     fi
   - which coveralls
@@ -155,10 +155,10 @@ install:
   - |
       if [[ $((ENDTIME - STARTTIME)) -lt 1200 ]]
       then
-        wget -q -nc https://zenodo.org/record/1169181/files/uxclumpy-cutoff.fits https://zenodo.org/record/1169181/files/uxclumpy-cutoff-omni.fits
-        export MODELDIR=$PWD
+        wget -q -nc https://zenodo.org/record/1169181/files/uxclumpy-cutoff.fits https://zenodo.org/record/1169181/files/uxclumpy-cutoff-omni.fits;
+        export MODELDIR=$PWD;
       else
-        echo "skipping model download, setup was slow"
+        echo "skipping model download, setup was slow";
       fi
   - popd
 
@@ -209,7 +209,7 @@ script:
       then
         PYTHONPATH=../../:${PYTHONPATH}:npyinterp/ LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:npyinterp/ bash runall.sh;
       else
-        echo "skipping tests, setup was slow"
+        echo "skipping tests, setup was slow";
       fi
       popd;
     fi
@@ -220,7 +220,7 @@ script:
       then
         PYTHONPATH=../../:$PYTHONPATH WITHAPEC=0 coverage run -p ../../examples/sherpa/xagnfitter.py;
       else
-        echo "skipping tests, setup was slow"
+        echo "skipping tests, setup was slow";
       fi
       popd;
     fi
