### 2021-03-08 JohannesBuchner/BXA#20 @JohannesBuchner
> Hi, some more comments on the example code:
>
>     * It seems that some of the sherpa examples rely on `npyinterp` and python 2? Could this be updated to python 3? Please correct me if I am wrong.

I think it should run with python3 (I just also fixed the test.py there). Can you send me an error message?

>     * The example code in `docker/` also seems a bit outdated (e.g. doesn't use `ultranest` and python 3). OK if you want to keep this for reference, but then an up-to-date docker example should be added too.
>     * What is the difference between `xagnfitter.py` detailed in the docs and `fitagn.py` in the docker examples?

The docker stuff is outdated. It is not mentioned in the paper and in the doc. It is only there in case someone wants to revive it.

Is it possible to not evaluate it, or do I have to delete it?

>     * When running `gal.py` I get a connection timeout to swift.ac.uk (looks likely to be a temporary problem on their side, so just to notify you)
>
>     * Can you fix the latest Travis CI build? Seems that the examples run fine and it should work

Yes, I think this is a small syntax issue in the CI code. Working on it.
