### 2021-03-08 637f661
address small doc issues from #19
