+++ README.rst
@@ -11,7 +11,6 @@ BXA provides the following features:
    * finding the best fit
    * computing error bars
    * computing marginal probability distributions
-   * parallelisation with MPI
 * plotting of spectral model vs. the data:
    * for the best fit
    * for each of the solutions (posterior samples)
@@ -29,14 +28,10 @@ BXA provides the following features:
 BXA shines especially
 
 * when systematically analysing a large data-set, or
-* when comparing multiple models
-* when analysing low counts data-set with realistic models
-
-because its robust and unsupervised fitting algorithm explores
-even complicated parameter spaces in an automated fashion.
-The user does not need to initialise to good starting points.
-The `algorithm <https://johannesbuchner.github.io/UltraNest/method.html>`_ automatically runs until convergence, and slows down to sample
-carefully if complicated parameter spaces are encountered. This allows building automated analysis pipelines.
+* when comparing multiple models, or
+* when analysing low counts data-set
+* when you don't want to babysit your fits
+* when you don't want to test MCMC chains for their convergence
 
 .. image:: https://img.shields.io/pypi/v/BXA.svg
         :target: https://pypi.python.org/pypi/BXA
@@ -76,7 +71,7 @@ BXA's `documentation <http://johannesbuchner.github.io/BXA/>`_ is hosted at http
 Installation
 -------------
 
-First, you need to have either `Sherpa`_ or `Xspec`_ installed and its environment loaded.
+First, you need to have `Sherpa`_ or `Xspec`_ installed and its environment loaded.
 
 BXA itself can installed easily using pip or conda::
 
@@ -97,8 +92,8 @@ You may need to install python and some basic packages through your package mana
 	$ yum install ipython python-matplotlib scipy numpy matplotlib
 	$ apt-get install python-numpy python-scipy python-matplotlib ipython
 
-BXA requires the following python packages: requests corner astropy h5py cython scipy tqdm.
-They should be downloaded automatically. If they are not, install them
+BXA requires the following python packages: cython tqdm h5py ultranest,
+which should be downloaded automatically. If they are not, install them
 also with pip/conda.
 
 The source code is available from https://github.com/JohannesBuchner/BXA,
@@ -111,14 +106,6 @@ so alternatively you can download and install it::
 Or if you only want to install it for the current user::
 
 	$ python setup.py install --user
-
-**Supported operating systems**: 
-BXA runs on all operating systems supported by 
-`ciao/sherpa <https://cxc.cfa.harvard.edu/ciao/watchout.html#install>`_ or 
-`heasoft/xspec <https://heasarc.gsfc.nasa.gov/lheasoft/issues.html>`_.
-The support is systematically tested for every BXA release by 
-`Travis CI <https://travis-ci.com/github/JohannesBuchner/BXA>`_, but only for Ubuntu Linux.
-
 
 Running
 --------------
@@ -143,8 +130,7 @@ For *Xspec*, start python or ipython::
 	
 	In [3]:	bxa.BXASolver?
 
-Now you can use BXA. See the documentation pages for how
-to perform analyses. Several examples are included.
+Now you can use BXA.
 
 .. _ultranest: http://johannesbuchner.github.io/UltraNest/
 
