### 2021-03-08 87f661f
address small doc issues from #19
