### 2021-03-08 2ba2ccf
added python package deps to install (#19)
