+++ README.rst
@@ -92,8 +92,8 @@ You may need to install python and some basic packages through your package mana
 	$ yum install ipython python-matplotlib scipy numpy matplotlib
 	$ apt-get install python-numpy python-scipy python-matplotlib ipython
 
-BXA requires the following python packages: cython tqdm h5py ultranest,
-which should be downloaded automatically. If they are not, install them
+BXA requires the following python packages: requests corner astropy h5py cython scipy tqdm.
+They should be downloaded automatically. If they are not, install them
 also with pip/conda.
 
 The source code is available from https://github.com/JohannesBuchner/BXA,
