+++ .travis.yml
@@ -31,7 +31,7 @@ before_cache:
 
 install:
   - sudo apt-get install -y python3-numpy python-numpy python3-dev python-tk python3-pip python3-setuptools
-  - sudo pip3 install --upgrade pip==20.3.4
+  - sudo pip3 install --upgrade pip
   - which python || true
   - which python3 || true
   - export MYCACHE=$HOME/Downloads/$TOOL
@@ -143,7 +143,7 @@ install:
 
   - |
     if [[ "$TOOL" != "sherpa-conda" ]]; then
-      sudo pip3 install requests "corner<2.2" astropy h5py cython coveralls coverage scipy tqdm;
+      sudo pip3 install requests corner astropy h5py cython coveralls coverage scipy tqdm;
       sudo pip3 install ultranest;
     fi
   - which coveralls
