+++ .travis.yml
@@ -3,11 +3,11 @@ language: generic
 os:
   - linux
 
-virtualenv:
-  system_site_packages: true
+#virtualenv:
+#  system_site_packages: true
 
-#python:
-#  - "3.6"
+python:
+  - "3.6"
 #  - "2.7"
 
 env:
