### 2021-03-08 8e29f55
[ci] avoid upgrading pip too high
