### 2021-03-08 a36f71b
[ci] load older pip for astropy
