+++ .travis.yml
@@ -31,7 +31,7 @@ before_cache:
 
 install:
   - sudo apt-get install -y python3-numpy python-numpy python3-dev python-tk python3-pip python3-setuptools
-#  - sudo pip3 install --upgrade pip
+  - sudo pip3 install --upgrade pip==20.3.4
   - which python || true
   - which python3 || true
   - export MYCACHE=$HOME/Downloads/$TOOL
