### 2021-03-09 b6ad061
add information on how to reduce verbosity
