### 2021-03-09 aae23ba
add information on how to reduce verbosity
