### 2021-03-09 JohannesBuchner/BXA#19 @JohannesBuchner
Hi,
>     * Installation: Add the full list of dependencies to the install steps. I also needed `cython`, `tqdm`, `h5py` and `ultranest` as a minimal list to get things running.
>     * In the "convenience features" documentation
>
>       * Seems the example call of `gal.py` is a typo as refers to another script

>       * Would be nice to have some quick code showing exactly how to import chains in Xspec or run a script in parallel

>     * The default ultranest output is cool, but could be distracting/annoying when running in certain settings. Could you show in the docs an easy interface in BXA to switch it off?

I believe I fixed these now.

>     * For both the sherpa/Xspec cases, it would be nice to see a full standalone demo including some example data. I know this is done in the scripts in `examples/` , but would be helpful if at least one case was treated notebook-style on the documentation pages.
>     * Similarly for the `xagnfitter.py` script - you have a case with some example data in the Travis CI, it could be nice to also have this in the docs as a walkthrough for users.

I will see what I can do here, unfortunately the sherpa and Xspec codebases are complex and out of my control, and make inline plotting very difficult. Perhaps I can place README files in the example folders which show the steps and the expected output. Would that address your point?
