### 2021-03-09 c32d131
[ci] install corner later, as it depends on hdf5 now
