### 2021-03-09 3a5cff1
[ci] add libhdf5-dev
