### 2021-03-09 43c8994
[ci] pin to older corner, because netcdf4 does not install
