### 2021-03-09 34b653a
[ci] command typo
