+++ .travis.yml
@@ -143,7 +143,7 @@ install:
 
   - |
     if [[ "$TOOL" != "sherpa-conda" ]]; then
-      sudo pip3 install requests corner<2.2 astropy h5py cython coveralls coverage scipy tqdm;
+      sudo pip3 install requests "corner<2.2" astropy h5py cython coveralls coverage scipy tqdm;
       sudo pip3 install ultranest;
     fi
   - which coveralls
