### 2021-03-09 JohannesBuchner/BXA#19 @cescalara
Great! Thanks, the updates look good.

> I will see what I can do here, unfortunately the sherpa and Xspec codebases are complex and out of my control, and make inline plotting very difficult. Perhaps I can place README files in the example folders which show the steps and the expected output. Would that address your point?

Ah, I see what you mean. I think a README solution as you propose would be fine. The main point I wanted to make is that it would be nice for users to have a quick check of their install against expected output that they can run out of the box in a guided way.
