### 2021-03-09 JohannesBuchner/BXA#20 @cescalara
Hi,

> I think it should run with python3 (I just also fixed the test.py there). Can you send me an error message?

Yes you're right, it runs fine. For me, the Makefile doesn't work out of the box and I had to use `np.get_include()` to grab the numpy headers path. Maybe you could include this in the Makefile or use a setup.py.

> The docker stuff is outdated. It is not mentioned in the paper and in the doc. It is only there in case someone wants to revive it.

OK, I understand, no need to delete it. I just got confused when wandering around the repo looking for examples as it is part of your CI tests. I was thinking it could be nice to have an up-to-date docker ready for users, given how it can be painful to install Xspec/Sherpa. However on second thoughts this would probably be a nightmare to maintain, so up to you if you want to add anything new.
