### 2021-03-09 JohannesBuchner/BXA#20 @cescalara
When checking the sherpa examples again I noticed that you need to `import os` in `example_pcabackground.py`. I guess this is a python 2/3 thing.
