### 2021-03-12 JohannesBuchner/BXA#20 @JohannesBuchner
> Yes you're right, it runs fine. For me, the Makefile doesn't work out of the box and I had to use `np.get_include()` to grab the numpy headers path. Maybe you could include this in the Makefile or use a setup.py.

Ideally, this would become a cython package with proper setup.py, but at the time I did not know how to do this. It is legacy code which only few people will need, which hopefully (like you) can adapt the Makefile if it doesn't work, and otherwise raise issues or email me.

> However on second thoughts this would probably be a nightmare to maintain,

Additionally, there are also licensing issues ... alas not all of it is FOSS.

> When checking the sherpa examples again I noticed that you need to import os, numpy in example_pcabackground.py. I guess this is a python 2/3 thing.

The script was written for "sherpa", which is a command launching an ipython environment with some modules preloaded. I added an import (6dd578cae), which hopefully enables use with the python3 command as well.

I hope this addresses the issues.
