### 2021-03-12 a1674ca
add doc for BXA/sherpa examples
