+++ examples/sherpa/README.rst
@@ -1,0 +1,667 @@
+BXA/Sherpa example scripts
+==========================
+
+This folder contains simple and complex examples
+how BXA can be invoked for spectral analysis.
+
+Please refer to https://johannesbuchner.github.io/BXA/
+for full documentation, including how to install BXA.
+
+
+Test data
+-------------------
+
+As test data, this includes the spectral file example-file.fak
+representing a ATHENA observation of an absorbed AGN, and the corresponding 
+response.
+
+Simple analysis
+-----------------
+
+Have a look at the file example_simplest.py. It contains:
+
+* Loading data
+* setting up a model and its parameter ranges
+* running a BXA fit with specified priors
+* plotting the posterior predictions (convolved with the response)
+* plotting the model (posterior predictions, not convolved)
+* making a Q-Q plot
+
+See https://johannesbuchner.github.io/BXA/ to understand the code.
+See https://johannesbuchner.github.io/UltraNest/ to understand the output of the
+fitting engine (for example, its `FAQ page <https://johannesbuchner.github.io/UltraNest/issues.html>`_).
+
+Expected output::
+
+	$ python3 example_simplest.py
+	read RMF file athenapp_ir_b4c_wfi_withfilter_fov40.0arcmin_avg.rsp
+	[ultranest] Sampling 400 live points from prior ...
+
+
+	Mono-modal Volume: ~exp(-4.18) * Expected Volume: exp(0.00) Quality: ok
+
+	mypow.PhoIndex:      +1.0|***********************************************************************************|     +3.0
+	mypow.norm    :  +1.0e-10|********** *** * ***   *** * * *    * *   *  *    *  *         *    *     *    *   | +1.0e+01
+
+	Z=-1751856.9(0.00%) | Like=-1618084.55..-4292.98 [-1.69e+08..-4793] | it/evals=80/487 eff=91.9540% N=400 0 
+
+	Mono-modal Volume: ~exp(-4.47) * Expected Volume: exp(-0.23) Quality: ok
+
+	mypow.PhoIndex:      +1.0|***********************************************************************************|     +3.0
+	mypow.norm    :  +1.0e-10|****************** *** *   * *  *  *   *   *****  **  **  *             *  +8.7e-02| +1.0e-01
+
+	Z=-35380.7(0.00%) | Like=-35235.95..-4292.98 [-1.69e+08..-4793] | it/evals=160/575 eff=91.4286% N=400 0 
+
+	Mono-modal Volume: ~exp(-4.47)   Expected Volume: exp(-0.45) Quality: ok
+
+	mypow.PhoIndex:      +1.0|***********************************************************************************|     +3.0
+	mypow.norm    :  +1.0e-10|****************    ***  +2.8e-03                                                  | +1.0e-02
+
+	Z=-27187.5(0.00%) | Like=-27051.37..-4292.98 [-1.69e+08..-4793] | it/evals=240/670 eff=88.8889% N=400 
+
+	Mono-modal Volume: ~exp(-4.72) * Expected Volume: exp(-0.67) Quality: ok
+
+	mypow.PhoIndex:      +1.0|**************************************************************************** ******|     +3.0
+	mypow.norm    :  +1.0e-10|****************    **  +2.6e-03                                                   | +1.0e-02
+
+	Z=-22433.1(0.00%) | Like=-22407.73..-4292.98 [-1.69e+08..-4793] | it/evals=320/764 eff=87.9121% N=400 
+
+	Mono-modal Volume: ~exp(-4.97) * Expected Volume: exp(-0.90) Quality: ok
+
[…]
