### 2021-03-12 45e3fea
add output to xagnfitter script as well
