+++ doc/xagnfitter.rst
@@ -29,6 +29,1626 @@ Features:
 I strongly recommend using the `xagnfitter <https://github.com/JohannesBuchner/BXA/blob/master/examples/sherpa/xagnfitter.py>`_
 script instead of hardness ratios.
 
+xagnfitter.py script
+--------------------
+
 It is included verbatim below:
 
 .. literalinclude:: ../examples/sherpa/xagnfitter.py
+
+Example run
+--------------------
+
+You can try running this against the AGN spectrum provided in the examples/sherpa/chandra/ folder.
+
+The output should look something like this::
+
+    $ WITHAPEC=0 MODELDIR=$HOME/Downloads/specmodels/ python3 ../xagnfitter.py 
+    read ARF file cdfs4Ms_179.arf
+    [sherpa.astro.io INFO]: read ARF file cdfs4Ms_179.arf
+    read RMF file cdfs4Ms_179.rmf
+    [sherpa.astro.io INFO]: read RMF file cdfs4Ms_179.rmf
+    read ARF (background) file cdfs4Ms_179.arf
+    [sherpa.astro.io INFO]: read ARF (background) file cdfs4Ms_179.arf
+    read RMF (background) file cdfs4Ms_179.rmf
+    [sherpa.astro.io INFO]: read RMF (background) file cdfs4Ms_179.rmf
+    read background file cdfs4Ms_179_bkg.pi
+    [sherpa.astro.io INFO]: read background file cdfs4Ms_179_bkg.pi
+     Solar Abundance Vector set to wilm:  Wilms, J., Allen, A. & McCray, R. ApJ 542 914 (2000) (abundances are set to zero for those elements not included in the paper).
+     Cross Section Table set to vern:  Verner, Ferland, Korista, and Yakovlev 1996
+    loading nH from 179.pi.nh (expecting something like 1e21 in there)
+    setting galactic nH to 0.0088 [units of 1e22/cm²]
+    combining components
+    linking parameters
+    setting redshift
+    creating priors
+    setting source and background model ...
+    [bxa.Fitter INFO]: PCAFitter(for ID=1)
+    [bxa.Fitter INFO]: loading PCA information from /home/user/bin/ciao-4.13/ots/lib/python3.7/site-packages/bxa/sherpa/background/chandra_1024.json
+    [bxa.Fitter INFO]: fitting background of ID=1 using PCA method
+    [bxa.Fitter INFO]: have 2521 background counts for deconvolution
+    [bxa.Fitter INFO]: fit: initial PCA decomposition: [ 3.40159007e+00 -6.29801294e-04  1.15351627e-02  8.72323308e-03
+      7.12865031e-03 -8.44044155e-03 -4.73346905e-03  5.83916607e-04
+     -4.79908423e-04 -9.18585670e-04 -1.87723042e-04]
+    [bxa.Fitter INFO]: fit: first full fit done
+    [bxa.Fitter INFO]: fit: parameters: [-0.24972366790501033, 0.1237617574156954, 1.619517445974407, -0.3466035071997995, -0.7134659455645368, 1.364537126802302, 0.10510218498377895, -0.41467820359992413, 0.9144121142234537, -1.144703704590311, -2.4736204436412694]
+    tbvabs Version 2.3
+    Cosmic absorption with grains and H2, modified from
+    Wilms, Allen, & McCray, 2000, ApJ 542, 914-924
+    Questions: Joern Wilms
+    joern.wilms@sternwarte.uni-erlangen.de
+    joern.wilms@fau.de
+
+    http://pulsar.sternwarte.uni-erlangen.de/wilms/research/tbabs/
+
+    PLEASE NOTICE:
+    To get the model described by the above paper
+    you will also have to set the abundances:
[…]
