### 2021-03-12 JohannesBuchner/BXA#19 @JohannesBuchner
I added README files in the two folders:

https://github.com/JohannesBuchner/BXA/tree/master/examples/xspec
https://github.com/JohannesBuchner/BXA/tree/master/examples/sherpa

and the output of xagnfitter here:
https://johannesbuchner.github.io/BXA/xagnfitter.html

I believe this addresses the issue.
