### 2021-03-12 JohannesBuchner/BXA#19 @cescalara
Ok, I was thinking more some key plots (e.g. input spectrum, fit corner plot) with some explanation of what the example does or so for a quick visual check. Is there any reason you don't want to include the plots themselves? Text output is thorough but not so intuitive.
