### 2021-03-12 JohannesBuchner/BXA#19 @JohannesBuchner
It's a good idea, I will add it.
