### 2021-03-12 f0e5aa1
improve doc for examples
