+++ examples/sherpa/example_pcabackground.py
@@ -1,115 +1,105 @@
 from __future__ import print_function
-try:
-	import os, numpy
-	from sherpa.astro.ui import *
-	import bxa.sherpa as bxa
-	bxa.default_logging()
-	print('loading background fitting module...')
-	from bxa.sherpa.background.pca import auto_background
+import os, numpy
+from sherpa.astro.ui import *
+import bxa.sherpa as bxa
+bxa.default_logging()
+print('loading background fitting module...')
+from bxa.sherpa.background.pca import auto_background
 
-	# BXA fully supports fitting multiple ids with the usual id=2, otherids=(3,4,5) 
-	#     parameters
-	id = 2
-	load_pha(id, 'swift/interval0pc.pi')
-	if get_rmf(id).energ_lo[0] == 0: get_rmf(id).energ_lo[0] = 0.001
-	if get_arf(id).energ_lo[0] == 0: get_arf(id).energ_lo[0] = 0.001
-	set_stat('cstat')
-	set_xsabund('wilm')
-	set_xsxsect('vern')
+# BXA fully supports fitting multiple ids with the usual id=2, otherids=(3,4,5) 
+#     parameters
+id = 2
+load_pha(id, 'swift/interval0pc.pi')
+set_stat('cstat')
+set_xsabund('wilm')
+set_xsxsect('vern')
 
-	# next we set up a source model.
-	#    with automatic milky way absorption
-	galabso = bxa.auto_galactic_absorption(id)
-	model = xszpowerlw.src * xszwabs.abso * galabso
-	src.norm.val = 1e-5
+# next we set up a source model.
+#    with automatic milky way absorption
+galabso = bxa.auto_galactic_absorption(id)
+model = xszpowerlw.src * xszwabs.abso * galabso
+src.norm.val = 1e-5
 
-	#    with automatic background from PCA
-	set_model(id, model)
-	convmodel = get_model(id)
-	bkg_model = auto_background(id)
-	#set_bkg_full_model(id, bkg_model)
-	set_full_model(id, convmodel + bkg_model*get_bkg_scale(id))
+#    with automatic background from PCA
+set_model(id, model)
+convmodel = get_model(id)
+bkg_model = auto_background(id)
+#set_bkg_full_model(id, bkg_model)
+set_full_model(id, convmodel + bkg_model*get_bkg_scale(id))
 
-	id2 = None
-	if os.path.exists('interval0wt.pi'):
-		id2 = 3
-		load_pha(id2, 'interval0wt.pi')
-		if get_rmf(id2).energ_lo[0] == 0: get_rmf(id2).energ_lo[0] = 0.001
-		if get_arf(id2).energ_lo[0] == 0: get_arf(id2).energ_lo[0] = 0.001
-		if get_bkg(id2).counts.sum() > 100:
-			set_model(id2, model)
-			convmodel2 = get_model(id2)
-			bkg_model2 = auto_background(id2)
-			#set_bkg_full_model(id2, bkg_model2)
-			set_full_model(id2, convmodel2 + bkg_model2*get_bkg_scale(id2))
-		else:
-			delete_data(id2)
-			id2 = None
+id2 = None
+if os.path.exists('interval0wt.pi'):
+	id2 = 3
+	load_pha(id2, 'interval0wt.pi')
+	if get_rmf(id2).energ_lo[0] == 0: get_rmf(id2).energ_lo[0] = 0.001
+	if get_arf(id2).energ_lo[0] == 0: get_arf(id2).energ_lo[0] = 0.001
[…]
