+++ examples/sherpa/example_automatic_background_model.py
@@ -1,7 +1,6 @@
 from __future__ import print_function
-
+import os, numpy
 from sherpa.astro.ui import *
-import numpy
 import bxa.sherpa as bxa
 bxa.default_logging()
 print('loading background fitting module...')
@@ -96,7 +95,5 @@ priors += [limited_19_24] # galnh
 priorfunction = bxa.create_prior_function(priors = priors)
 print('running BXA ...')
 solver = bxa.BXASolver(id, prior=priorfunction, parameters = parameters, 
-	outputfiles_basename = 'superfit_')
+	outputfiles_basename = 'superfit')
 solver.run(resume=True)
-
-exit()
