+++ examples/sherpa/runall.sh
@@ -9,4 +9,7 @@ coverage run -p example_automatic_background_model.py || exit 3
 rm testmodel.npz  && coverage run -p example_rebinning.py || exit 4
 coverage run -p example_rebinning.py || exit 5
 
-coverage run -p model_compare.py superfit_ wabs_noz_/
+coverage run -p model_compare.py superfit/ tbabs_noz/
+
+mkdir -p reference-output
+convert simplest/plots/corner.pdf reference-output/corner.png
