+++ paper.bib
@@ -170,3 +170,12 @@
   owner = {user},
   timestamp = {2013.09.02}
 }
+
+@misc{pyblocxs,
+  author = {{CHASC Astro-Statistics Collaboration}},
+  title = {{pyblocxs}},
+  year = {2011},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/brefsdal/pyblocxs}
+}
