### 2021-03-12 #3045 @JohannesBuchner
> And also a few comments on the paper:
>
>     * What is the relationship to [pyblocxs](https://github.com/brefsdal/pyblocxs)? I realise that this package is no longer active, and little of the original code is left. however, if this code contributed to the ideas and development of BXA, it seems proper to note this and acknowledge the author in the spirit of good open-source etiquette.

Good idea. I used their code to understand how to hook into sherpa, but as you say, little of that is left. I added a line in the acknowledgements.

>     * For me, it would be useful to have a high-level description of Xspec/sherpa and their differences in the paper. It would also make sense to mention [3ML](https://github.com/threeML/threeML) as an alternative tool for X-ray analysis and the differences there.

There are a bunch of environments, including Spex, ISIS, 3ML, Xspec, sherpa, and probably others (it is hard to do a full survey). It is possible to do meaningful spectral analysis with any of them, they do not differ fundamentally (only in defaults, bells and whistles, models and programming language).

>     * Would be nice to have a short overview of the methods/functionality in the paper in addition to pointing to your 2014 paper for details.

I am a bit unsure if it is possible to explain all of the concepts (Evidence estimation, Bayes factors, dealing with posterior chains) and their caveats within the expected word limit. I added some text that describes what functionality is included (taken from the readme file). I hope this addresses this.

> I really hope my comments are helpful and I'm happy to discuss on any of the points.

They are!
