### 2021-03-12 39a09b9
add ref to MCMC prior work
