+++ paper.bib
@@ -179,3 +179,23 @@
   journal = {GitHub repository},
   url = {https://github.com/brefsdal/pyblocxs}
 }
+@ARTICLE{Dyk2001,
+  author = {{van Dyk}, D.~A. and {Connors}, A. and {Kashyap}, V.~L. and {Siemiginowska},
+	A.},
+  title = {{Analysis of Energy Spectra with Low Photon Counts via Bayesian Posterior
+	Simulation}},
+  journal = {\apj},
+  year = {2001},
+  volume = {548},
+  pages = {224-243},
+  month = feb,
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+  adsurl = {http://adsabs.harvard.edu/abs/2001ApJ...548..224V},
+  doi = {10.1086/318656},
+  eprint = {arXiv:astro-ph/0008170},
+  file = {Published version:2001ApJ...548..224V.pdf:PDF;arXiv v1:2001ApJ...548..224V-eprintv1.pdf:PDF},
+  keywords = {Methods: Data Analysis, Methods: Statistical},
+  owner = {user},
+  review = {Bayesian X-ray spectral analysis using MCMC},
+  timestamp = {2012.08.26}
+}
