### 2021-03-12 #3045 @JohannesBuchner
@jgostick I read the COI guidelines, and wanted to mention that @grburgess is employed at the same institute as me. None of the other COI criteria apply, however (no joint papers, funded work, etc.). Not sure if it is a real issue, but I wanted to bring it to your attention for the sake of transparency.
