### 2021-03-13 #3045 @cescalara
> There are a bunch of environments, including Spex, ISIS, 3ML, Xspec, sherpa, and probably others (it is hard to do a full survey). It is possible to do meaningful spectral analysis with any of them, they do not differ fundamentally (only in defaults, bells and whistles, models and programming language).

Sure, A full survey of what's out there isn't easy. Perhaps just list the tools you mentioned (I wasn't aware of Spex/ISIS) without going into details, and highlight what separates BXA. As far as I can tell, BXA is cool as it is an interface to Bayesian analysis tools that you can easily apply to existing analyses in the very widely used Xspec and sherpa frameworks. So, you get lots of nice features without having to learn a whole new framework. I'm mainly asking this to check off the "State of the field: Do the authors describe how this software compares to other commonly-used packages?" box.

> I am a bit unsure if it is possible to explain all of the concepts (Evidence estimation, Bayes factors, dealing with posterior chains) and their caveats within the expected word limit. I added some text that describes what functionality is included (taken from the readme file). I hope this addresses this.

Yup, looks good. It is nice to have an overview of the main functionality.
