### 2021-03-13 #3045 @grburgess
At the moment, I am unable to get  Sherpa to install in a clean virtual environment with python 3.8 on MacOS.

And pip.

It is also not possible to install XSPEC in a non system-breaking way with  homebrew.

@cescalara how were you able to work around this?
