### 2021-03-14 #3045 @grburgess
Are there notes somewhere for how to install Sherpa + CIAO on Mac OS? @cescalara which platforms did you use?
