### 2021-03-14 #3045 @cescalara
Hi @grburgess,

I tested the installation on Ubuntu in a docker env and followed the instructions linked to in the docs.
* CIAO/Sherpa: https://cxc.harvard.edu/ciao/download/
* Xspec: https://heasarc.gsfc.nasa.gov/

I did not yet have time to test on MacOS, but I want to do so.

Some comments for @JohannesBuchner:
* I found the whole Sherpa/CIAO part a bit confusing, it seems there is also a python-only version of Sherpa that can be pip-installed, but this doesn't have the functionality required to run BXA with Xspec models. Can you clarify the install steps and if CIAO is also needed?
* I didn't raise this before as what was present satisfied the basic requirements of the JOSS review criteria (quoted below). However, given how Xspec and/or Sherpa are such important dependencies for BXA and can be difficult to install, I think it would make sense for BXA to come with `install_xspec` and `install_sherpa` scripts for the user to get up and running.

JOSS review criteria for reference:
> Installation instructions
>
>There should be a clearly-stated list of dependencies. Ideally these should be handled with an automated package management solution.
>
>Good: A package management file such as a Gemfile or package.json or equivalent
>OK: A list of dependencies to install
>Bad (not acceptable): Reliance on other software not listed by the authors
