### 2021-03-14 #3045 @grburgess
@cescalara aha. Ok, I am currently also installing everything in a linux environment so I hope to be able to proceed from there. I have experience with XSPEC in both linux and macos environments and will only be able to test the XSPEC version within linux. XSPEC now has a docker... but this of course means linux only.

I agree, the sherpa install is very confusing and the capabilities required by BXA via sherpa are unclear at the moment from the documentation. Granted, users familiar with either environment may have the experience to circumvent these issues, it should be documented a bit more clearly what exactly is needed.

Perhaps the pip install could perform some kind of system probe to tell the users what capabilities/features BXA can utilize with their current setup?
