### 2021-03-14 #3045 @cescalara
> Perhaps the pip install could perform some kind of system probe to tell the users what capabilities/features BXA can utilize with their current setup?

This would certainly be useful.
