### 2021-03-14 #3045 @JohannesBuchner
There are four possibilities to have a working ciao/sherpa environment:
- install with their conda channel: https://cxc.cfa.harvard.edu/ciao/download/
- install with the ciao-install installation bash script https://cxc.cfa.harvard.edu/ciao/download/ciao_install.html
- install only sherpa through pip (and optionally obtain xspec models manually)
- use a server at your institute where ciao/sherpa is already installed.

Please understand that these are complex programs that are outside of my control. They use non-standard installation methods, and they keep changing every year (and break occasionally). There is no way for me to provide up-to-date installation instructions, or automated installation script. Packaging ciao or heasoft into a docker container or conda package leads to legal problems, as not all of what is contained is FOSS. These options would require me to become a maintainer of these indefinitely and provide trouble-shooting of packages without being able to control the release process.

To give an idea of the difficulty, the heasoft/xspec download file changes URL with every new version, and there is no option other than parsing the HTML of the web page to find out what version is current. Conda is its own nightmare, with environments sometimes unable to resolve dependencies, preventing installation of further packages.

The assumption is that users already use pyxspec or sherpa, and only complement this with BXA, which has the role of a plugin.

The xspec models are not required for BXA with sherpa, although the examples use them.

> Perhaps the pip install could perform some kind of system probe to tell the users what capabilities/features BXA can utilize with their current setup?

I am not sure what is meant here. If you have sherpa (as tested by running the `sherpa` command), you can use [bxa.sherpa](https://johannesbuchner.github.io/BXA/sherpa-analysis.html), if you have pyxspec (as tested with `import xspec` in python), you can use [bxa.xspec](https://johannesbuchner.github.io/BXA/xspec-analysis.html).

I am certainly happy to make it as easy as possible for users (improve install docs, provide tools). But I am unsure how I can do it in a sustainable manner.
