### 2021-03-14 JohannesBuchner/BXA#21 @grburgess
JOSS review: CI and testing  
Is there a link to the CI tests that can be included in README showing which software platforms BXA currently runs on (eg. linux/macos) as well the python versions?

These could be useful for building examples that work out-of-the-box as well as informing the users of pitfalls in installing the software.
