### 2021-03-14 JohannesBuchner/BXA#22 @grburgess
JOSS review: installation 
Here are the issues with getting into a state of being able to run the ```python example_simplest.py ```

1) The main required dependency for the fitting, ```ultranest``` is not installed automatically either by pip or python setup.py install

2) The setup currently does not run on MacOS.

All of the following is run in a clean virtual environment on **Ubuntu 20.04.2** with **python 3.8** (no conda)

After completing the install for CIAO with the following steps:

```bash

bash ciao_install
pip install sherpa # (installs numpy)
python setup.py install # inside BXA folder cloned from GitHub

```

the following steps need to be run to avoid errors:

```bash

pip install ultranest

```

From this point I try to run ```python example_simplest.py ```

Result:

```
WARNING: imaging routines will not be available,
failed to import sherpa.image.ds9_backend due to
'RuntimeErr: DS9Win unusable: Could not find ds9 on your PATH'
WARNING: failed to import sherpa.astro.xspec; XSPEC models will not be available
read RMF file athenapp_ir_b4c_wfi_withfilter_fov40.0arcmin_avg.rsp
Traceback (most recent call last):
  File "example_simplest.py", line 20, in <module>
    set_model(xspowerlaw.mypow)
NameError: name 'xspowerlaw' is not defined
```

It would be useful to modify the following:

- [ ] add all required dependencies to the install/setup script

- [x] include examples that work with Sherpa alone as it is the easiest entry point into testing the code.

- [x]  as BXA is a plugin to the XSPEC/Sherpa analysis software, warning to the user upon installation that the environment does not detect the required essential dependencies  to run upon install.

- [x] perhaps an end-to-end example in the documentation (including data) would be helpful.

While it is not BXA should not be responsible for upkeep of XSPEC/Sherpa's installation procedures and capabilities, its focus as a plugin to these software means that it needs to perhaps guide the user thru the installation of its core functionality.

In regards to the software not installing on MacOS, this is again an issue of the core packages, but should be reflected in the caveats / platforms listed in the README and documentation.  A few of these may be addressed with https://github.com/JohannesBuchner/BXA/issues/21
