### 2021-03-14 #3045 @grburgess
I've listed a few issues with getting started here:

https://github.com/JohannesBuchner/BXA/issues/21
https://github.com/JohannesBuchner/BXA/issues/22

Currently, running examples is available only on linux and require XSPEC. Thus testing a Sherpa only install appears impossible, but I am perhaps missing something.
