### 2021-03-14 JohannesBuchner/BXA#22 @grburgess
The XSPEC install works perfectly. However, there are a number of required packages that not installed with either the ```ultranest``` package or BXA. These include:

```scipy```
```astropy```
```tqdm```
```h5py```

can these be added to the requirements so that the examples work out-of-the-box?
