### 2021-03-14 JohannesBuchner/BXA#23 @grburgess
JOSS Review: XSPEC Examples output
The gen.xspec script functions properly and the examples all run correctly.

It would be nice if the large amount of output from the sampler was removed from the docs / README as it is difficult to find the relevant information without a lot of scrolling. Is the output needed for any purpose?
