### 2021-03-14 JohannesBuchner/BXA#24 @grburgess
JOSS review: priors / transformations
The documentation refers to "prior transformations" throughout the docs. This is a reference to transforming the nested sampling focus on transforming all priors from a unit cube.

This could be confusing as users who do not realize that this is specifically for nested sampling and not a form of MCMC that is provided in both XSPEC (https://asd.gsfc.nasa.gov/XSPECwiki/Markov_Chain_Monte_Carlo) and Sherpa (https://sherpa.readthedocs.io/en/latest/mcmc/index.html). It might be useful at the introduction (since it is not in the name) to clarify this? The connection to nested sampling is clearly there, but a note or something similar could be more instructive.

Perhaps is would useful to define what a transformation is here:
https://github.com/JohannesBuchner/BXA/blob/master/bxa/xspec/solver.py#L104

**There are references to pymultinest throughout the docs.**

What are all the available priors and can a user add custom priors?

In the examples, a Gaussian prior must be specified with limits. Is it not possible to have an unbounded Gaussian prior?

It seems there is a wealth of functionality with priors in the examples that is perhaps not documented and a naive user could miss these capabilities.
