### 2021-03-14 JohannesBuchner/BXA#25 @grburgess
JOSS review: Documentation
In the documentation several important features are listed but without explanation. It would be helpful if these were explained quantitatively with examples / comparisons:

> when systematically analysing a large data-set

How does this differ from what is available in XSPEC/Sherpa?

> when comparing multiple models

This is explained well for those initiated with Bayes factor model comparison, but a link to that part of the docs would be useful to make the point

> when analysing low counts data-set

This claim is made [here](https://sherpa.readthedocs.io/en/latest/mcmc/index.html) and [here](https://hea-www.harvard.edu/astrostat/pyblocxs/) as well as the normal use of proper Poisson likelihoods. What specific features makes BXA shine in comparison? Can links be provided as well as comparative examples?

> when you don’t want to babysit your fits

What is meant by this statement? As stated above, can examples be linked/ shown that validate this?
I'm assuming it is because there is a stopping criteria in nested sampling, but is this meant to inform the user that they do not need to access the correctness of their posterior?

> when you don’t want to test MCMC chains for their convergence

I suppose this could be answered with the previous question.

In general, it could be important to fully separate what is a feature or strength of BXA vs what is inherited from the package it relies on / compliments.
