### 2021-03-14 #3045 @grburgess
As the XSPEC install / examples are working on linux I've completed the basic review of this part of the code and documentation:

[#23](https://github.com/JohannesBuchner/BXA/issues/23)
[#24](https://github.com/JohannesBuchner/BXA/issues/24)
[#25](https://github.com/JohannesBuchner/BXA/issues/25)
