### 2021-03-14 #3045 @grburgess
@JohannesBuchner

> Please understand that these are complex programs that are outside of my control. They use non-standard installation methods, and they keep changing every year (and break occasionally). There is no way for me to provide up-to-date installation instructions, or automated installation script. Packaging ciao or heasoft into a docker container or conda package leads to legal problems, as not all of what is contained is FOSS. These options would require me to become a maintainer of these indefinitely and provide trouble-shooting of packages without being able to control the release process.

>To give an idea of the difficulty, the heasoft/xspec download file changes URL with every new version, and there is no option other than parsing the HTML of the web page to find out what version is current. Conda is its own nightmare, with environments sometimes unable to resolve dependencies, preventing installation of further packages.

It is understood, and there is no way to put the burden of maintenance of an enhancement software. Have you communicated with the maintainers of these softwares for the possibility of releasing dockers or some other work around? I've seen statements by the XSPEC maintainers that hint at them not caring too much what is done with the code. But you are providing an enhancement and service to their codes which would multiply their user base. So it would seem logical that they could work with you to ease the access to new users. Of course Conda is always an option and that eases many of the issues. Perhaps (as stated in some of the issues) it is better to inform users of BXA what the current platform limitations are?

> I am not sure what is meant here. If you have sherpa (as tested by running the sherpa command), you can use bxa.sherpa, if you have pyxspec (as tested with import xspec in python), you can use bxa.xspec.

In the documentation you have stated to first have the required analysis software installed. Perhaps you can run a script upon install that does some test imports and loudly informs the user that XXX software is not available and will BXA will not function without them. Of course, an experienced user can and will know this ahead of time, I'm more concerned with e.g. a student who wants to do nested sampling, finds BXA and gets frustrated with things not working on the first try. BXA does a good job warning the user about the fitting process:

```bash
Null hypothesis probability of 0.000000e+00 with 795 degrees of freedom
 Current data and model not fit yet.
  jeffreys prior for nH between 1.000000e-02 and 1.000000e+03
  jeffreys prior for norm between 1.000000e-10 and 1.000000e+01
   note: this parameter spans *many* dex. Double-check the limits are reasonable.
```

I think it would ease some pain by being verbose on the install.
