### 2021-03-14 96c9e48
fix requirements & provide user warning if neither xspec nor sherpa are there
