### 2021-03-14 7596fc6
allow using sherpa without xspec models in simplest example
