+++ examples/sherpa/README.rst
@@ -39,596 +39,596 @@ Expected output::
 	[ultranest] Sampling 400 live points from prior ...
 
 
-	Mono-modal Volume: ~exp(-4.18) * Expected Volume: exp(0.00) Quality: ok
+	Mono-modal Volume: ~exp(-4.69) * Expected Volume: exp(0.00) Quality: ok
 
-	mypow.PhoIndex:      +1.0|***********************************************************************************|     +3.0
-	mypow.norm    :  +1.0e-10|********** *** * ***   *** * * *    * *   *  *    *  *         *    *     *    *   | +1.0e+01
+	mypow.gamma:      +1.0|************************************************************************************|     +3.0
+	mypow.ampl :  +1.0e-10|************ ** * * *  ***    * * * * *  *      * * *    * *        *     **  *   * | +1.0e+01
 
-	Z=-1751856.9(0.00%) | Like=-1618084.55..-4292.98 [-1.69e+08..-4793] | it/evals=80/487 eff=91.9540% N=400 0 
+	Z=-1436853.7(0.00%) | Like=-1425796.15..-4079.21 [-1.09e+08..-4333] | it/evals=80/481 eff=98.7654% N=400 0 
 
-	Mono-modal Volume: ~exp(-4.47) * Expected Volume: exp(-0.23) Quality: ok
+	Mono-modal Volume: ~exp(-4.69)   Expected Volume: exp(-0.23) Quality: ok
 
-	mypow.PhoIndex:      +1.0|***********************************************************************************|     +3.0
-	mypow.norm    :  +1.0e-10|****************** *** *   * *  *  *   *   *****  **  **  *             *  +8.7e-02| +1.0e-01
+	mypow.gamma:      +1.0|************************************************************************************|     +3.0
+	mypow.ampl :  +1.0e-10|*********** *** ***  * *** ***   **        **  *      *      * ** *  +7.9e-02       | +1.0e-01
 
-	Z=-35380.7(0.00%) | Like=-35235.95..-4292.98 [-1.69e+08..-4793] | it/evals=160/575 eff=91.4286% N=400 0 
+	Z=-49990.0(0.00%) | Like=-48482.70..-4079.21 [-1.09e+08..-4333] | it/evals=160/566 eff=96.3855% N=400 0 
 
-	Mono-modal Volume: ~exp(-4.47)   Expected Volume: exp(-0.45) Quality: ok
+	Mono-modal Volume: ~exp(-4.69)   Expected Volume: exp(-0.45) Quality: ok
 
-	mypow.PhoIndex:      +1.0|***********************************************************************************|     +3.0
-	mypow.norm    :  +1.0e-10|****************    ***  +2.8e-03                                                  | +1.0e-02
+	mypow.gamma:      +1.0|************************************************************************************|     +3.0
+	mypow.ampl :  +1.0e-10|************* ***   *   * *  +3.2e-03                                               | +1.0e-02
 
-	Z=-27187.5(0.00%) | Like=-27051.37..-4292.98 [-1.69e+08..-4793] | it/evals=240/670 eff=88.8889% N=400 
+	Z=-28599.1(0.00%) | Like=-28584.11..-4079.21 [-1.09e+08..-4333] | it/evals=240/650 eff=96.0000% N=400 
 
-	Mono-modal Volume: ~exp(-4.72) * Expected Volume: exp(-0.67) Quality: ok
+	Mono-modal Volume: ~exp(-5.23) * Expected Volume: exp(-0.67) Quality: ok
 
-	mypow.PhoIndex:      +1.0|**************************************************************************** ******|     +3.0
-	mypow.norm    :  +1.0e-10|****************    **  +2.6e-03                                                   | +1.0e-02
[…]
