### 2021-03-14 JohannesBuchner/BXA#22 @JohannesBuchner
> After completing the install for CIAO with the following steps:
>
> ```shell
> bash ciao_install
> pip install sherpa # (installs numpy)
> python setup.py install # inside BXA folder cloned from GitHub
> ```

You should not do `pip install sherpa`. ciao comes with sherpa and the xspec models. I suspect that the pip install overwrote something. Please follow the steps on https://cxc.cfa.harvard.edu/ciao/download/ciao_install.html

> the following steps need to be run to avoid errors:
>
> ```shell
> pip install ultranest
> ```
>
> From this point I try to run `python example_simplest.py `
>
> Result:
>
> ```
> WARNING: imaging routines will not be available,
> failed to import sherpa.image.ds9_backend due to
> 'RuntimeErr: DS9Win unusable: Could not find ds9 on your PATH'
> WARNING: failed to import sherpa.astro.xspec; XSPEC models will not be available
> read RMF file athenapp_ir_b4c_wfi_withfilter_fov40.0arcmin_avg.rsp
> Traceback (most recent call last):
>   File "example_simplest.py", line 20, in <module>
>     set_model(xspowerlaw.mypow)
> NameError: name 'xspowerlaw' is not defined
> ```
>
> It would be useful to modify the following:
>
>     * [ ]  add all required dependencies to the install/setup script

done, and release 4.0.3 includes them.

>     * [ ]  include examples that work with Sherpa alone as it is the easiest entry point into testing the code.

done, modified example_simplest.py

>     * [ ]   as BXA is a plugin to the XSPEC/Sherpa analysis software, warning to the user upon installation that the environment does not detect the required essential dependencies  to run upon install.

added.

>     * [ ]  perhaps an end-to-end example in the documentation (including data) would be helpful.

solved in #20 now, available at https://github.com/JohannesBuchner/BXA/tree/master/examples/sherpa as README.
