### 2021-03-14 f487424
add more setup info
