### 2021-03-14 JohannesBuchner/BXA#21 @JohannesBuchner
BXA does not contain any platform-dependent code, so it is only a matter of whether xspec/sherpa runs on that platform.

CI capabilities are very limited, because installation takes over 1 hour.

BXA does not have any strong python version requirements, python 2.7 is still supported, as are early 3 versions. If not, it is a bug. However, these days it is difficult to create a python 2.7 environment with the dependencies (e.g., astropy).
