### 2021-03-14 JohannesBuchner/BXA#23 @JohannesBuchner
I thought it would be instructive to see the output and the progress evolution while scrolling through.
