### 2021-03-15 JohannesBuchner/BXA#22 @grburgess
setup.py is still missing SciPy. Other than that, XSPEC examples run out-of-box on linux
