+++ examples/sherpa/swift/fit.py
@@ -1,4 +1,9 @@
 from __future__ import print_function
+from sherpa.astro.ui import *
+import numpy as np
+import os
+import json
+import numpy
 import bxa.sherpa as bxa
 print('loading background fitting module...')
 from bxa.sherpa.background.models import SwiftXRTBackground
@@ -14,8 +19,7 @@ logging.getLogger().addHandler(consoleHandler)
 
 id = 2
 load_pha(id, 'interval0pc.pi')
-if get_rmf(id).energ_lo[0] == 0: get_rmf(id).energ_lo[0] = 0.01
-if get_arf(id).energ_lo[0] == 0: get_arf(id).energ_lo[0] = 0.01
+nhgal = np.loadtxt('interval0pc.pi.nh')
 set_xlog()
 set_ylog()
 set_stat('cstat')
@@ -59,25 +63,10 @@ ignore(None, 0.5)
 ignore(5, None)
 notice(0.5, 5)
 
-import json
-props = json.load(open('index.json'))
-
-redshift_uncertain = False
-if not (props['z'] > 0) or os.environ.get('USEZ', '1') == '0':
-	redshift_uncertain = True
-	print(('no redshift in index.json', props['z']))
-	#exit()
-	#assert False, ('wrong redshift in index.json', props['z'])
-if 'nhgaltot' in props:
-	props['nhgal'] = props['nhgaltot']
-if not 1e18 < props['nhgal'] < 1e24:
-	print(('wrong galNH in index.json', props['nhgal']))
-	exit()
-	assert False, ('wrong galNH in index.json', props['nhgal'])
-
 #from bxa.sherpa.rebinnedmodel import RebinnedModel
-load_table_model('sphere', '/opt/models/sphere0708.fits')
-load_table_model('sphere2', '/opt/models/sphere0708.fits')
+# sphere model of Brightman & Nandra (2011)
+load_table_model('sphere', os.environ.get('MODELDIR', '.') + '/sphere0708.fits')
+load_table_model('sphere2', os.environ.get('MODELDIR', '.') + '/sphere0708.fits')
 
 srcmodel = sphere * xstbabs.galabso
 srcmodel2 = sphere2 * xstbabs.galabso
@@ -91,35 +80,35 @@ sphere.phoindex.min = 1
 sphere.phoindex.max = 3
 sphere.phoindex.val = 2
 sphere.norm.min = 1e-10
-sphere.norm.max = 100
+sphere.norm.max = 1000
 sphere.norm.val = 0.001
 sphere.redshift.max = 10
 sphere2.phoindex.min = 1
 sphere2.phoindex.max = 3
 sphere2.phoindex.val = 2
 sphere2.norm.min = 1e-10
-sphere2.norm.max = 100
+sphere2.norm.max = 1000
 sphere2.norm.val = 0.001
 sphere2.redshift.max = 10
 from sherpa.models.parameter import Parameter
-redshift = Parameter('src', 'z', 2 if redshift_uncertain else props['z'], 0, 10, 0, 10)
+redshift = 2.352
 sphere.redshift = redshift
-sphere2.redshift = redshift
+sphere2.redshift = sphere.redshift
 # assuming solar abundances ...
 sphere.abund_.freeze()
 sphere.fe_abund_.freeze()
 sphere.redshift = redshift
-sphere.nH.min = 1e20 / 1e22
-sphere.nH.max = 1e26 / 1e22
-sphere.nH.val = 1e22 / 1e22
+sphere.nh.min = 1e20 / 1e22
+sphere.nh.max = 1e26 / 1e22
+sphere.nh.val = 1e22 / 1e22
 sphere2.abund_.freeze()
 sphere2.fe_abund_.freeze()
[…]
