### 2021-03-29 644082b
replace mentions of multinest; make likelihood and priors accessible
