### 2021-03-29 2c9d037
add warning to end of install as well
