### 2021-03-29 16c0cbe
document how prior transforms work better
