### 2021-03-29 15083a1
remove long output to address. Closes #23.
