+++ examples/xspec/README.rst
@@ -120,551 +120,9 @@ Expected output::
 
 	Z=-24616.1(0.00%) | Like=-24611.37..-4277.72 [-1.045e+08..-4464] | it/evals=160/9998 eff=inf% N=400 0 
 
-	Mono-modal Volume: ~exp(-4.71) * Expected Volume: exp(-0.45) Quality: correlation length: 3 (+)
-
-	PhoIndex :      +1.0|************************************************  ******************************************************|     +3.0
-	log(norm):     -10.0|***********************************************************************  -2.6                           |     +1.0
-
-	Z=-24587.0(0.00%) | Like=-24580.29..-4277.72 [-1.045e+08..-4464] | it/evals=240/9998 eff=inf% N=400 
-
-	Mono-modal Volume: ~exp(-5.09) * Expected Volume: exp(-0.67) Quality: correlation length: 3 (+)
-
-	PhoIndex :      +1.0|******************************************** ***  ********************************************** *******|     +3.0
-	log(norm):     -10.0|       -8.6  **********************************************************  -2.6                           |     +1.0
-
-	Z=-22026.7(0.00%) | Like=-22012.60..-4249.81 [-1.045e+08..-4464] | it/evals=320/9998 eff=inf% N=400 
-
-	Mono-modal Volume: ~exp(-5.22) * Expected Volume: exp(-0.90) Quality: correlation length: 3 (+)
-
-	PhoIndex :      +1.0|************************************************  ********************************************** *******|     +3.0
-	log(norm):     -10.0|                  -7.4  * ********************************************  -2.7                            |     +1.0
-
-	Z=-17640.5(0.00%) | Like=-17606.95..-4249.81 [-1.045e+08..-4464] | it/evals=440/9998 eff=inf% N=400 
-
-	Mono-modal Volume: ~exp(-5.62) * Expected Volume: exp(-1.12) Quality: correlation length: 7 (-)
-
-	PhoIndex :      +1.0|************* *********** *********************************************** ********************** *******|     +3.0
-	log(norm):     -10.0|                       -6.9  ****************************************  -2.7                             |     +1.0
-
-	Z=-14906.6(0.00%) | Like=-14871.21..-4107.51 [-1.045e+08..-4464] | it/evals=520/9998 eff=inf% N=400 
-
-	Mono-modal Volume: ~exp(-5.62)   Expected Volume: exp(-1.35) Quality: correlation length: 7 (-)
-
-	PhoIndex :      +1.0|************* ******************************** **************************  ***************** *** *******|     +3.0
-	log(norm):     -10.0|                             -6.2  *********************************  -2.8                              |     +1.0
-
-	Z=-12616.7(0.00%) | Like=-12563.04..-4107.51 [-1.045e+08..-4464] | it/evals=600/9998 eff=inf% N=400 
-
-	Mono-modal Volume: ~exp(-5.62)   Expected Volume: exp(-1.57) Quality: correlation length: 138 (+)
-
-	PhoIndex :      +1.0|************************************************************************* ****************** *** *******|     +3.0
-	log(norm):     -10.0|                                 -5.8  *****************************  -2.9                              |     +1.0
-
-	Z=-11330.7(0.00%) | Like=-11317.59..-4107.51 [-1.045e+08..-4464] | it/evals=680/9998 eff=inf% N=400 
-
-	Mono-modal Volume: ~exp(-5.62)   Expected Volume: exp(-1.80) Quality: correlation length: 138 (+)
-
-	PhoIndex :      +1.0|************************************************************************* ****** *********** *** **  ** |     +3.0
-	log(norm):     -10.0|                                    -5.5  *************************  -3.0                               |     +1.0
-
[…]
