### 2021-03-29 JohannesBuchner/BXA#22 @JohannesBuchner
BXA only requires scipy for some plotting niceties for xspec. However, sometimes scipy is difficult to install on servers, so I do not want to make it a formal requirement in setup.py.

I updated the doc to recommend the installation of scipy, and mention in the doc which functionality requires scipy.

Let me know if there are any issues remaining here.
