### 2021-03-29 #3045 @JohannesBuchner
> It is understood, and there is no way to put the burden of maintenance of an enhancement software. Have you communicated with the maintainers of these softwares for the possibility of releasing dockers or some other work around?

I have (with both packages), but without success.

> > I am not sure what is meant here. If you have sherpa (as tested by running the sherpa command), you can use bxa.sherpa, if you have pyxspec (as tested with import xspec in python), you can use bxa.xspec.
>
> In the documentation you have stated to first have the required analysis software installed. Perhaps you can run a script upon install that does some test imports and loudly informs the user that XXX software is not available and will BXA will not function without them.
> I think it would ease some pain by being verbose on the install.

Added by commit https://github.com/JohannesBuchner/BXA/commit/2c9d037827e6e837e24656699dafd49a9471074f (included in v4.0.4)
