### 2021-03-29 2f5e6e2
remove duplicate examples
