+++ examples/sherpa/xz_auto.py
@@ -1,262 +1,0 @@
-"""
-
-Fit a X-ray point source (e.g. AGN), potentially obscured with BXA
-
-
-How to run this script:
-
-1) enter the docker container:
-
-$ xhost +local:`docker inspect --format='{{ .Config.Hostname }}' johannesbuchner/bxa_absorbed`
-$ docker run -v  /tmp/.X11-unix:/tmp/.X11-unix -e DISPLAY=$DISPLAY -v $PWD:/opt/example/ -ti johannesbuchner/bxa_absorbed bash
-
-2) load ciao and run the script:
-
-root@7082a4c30d82:/opt/example# . /opt/ciao-4.8/bin/ciao.sh
-root@7082a4c30d82:/opt/example# FILENAME=combined_src.pi ELO=0.5 EHI=8 sherpa fitagn.py
-
-"""
-from sherpa.astro.ui import *
-import os
-import sys
-import numpy
-
-import bxa.sherpa as bxa
-from bxa.sherpa.background.pca import auto_background
-from sherpa.models.parameter import Parameter
-
-import logging
-logging.basicConfig(filename='bxa.log',level=logging.DEBUG)
-logFormatter = logging.Formatter("[%(name)s %(levelname)s]: %(message)s")
-consoleHandler = logging.StreamHandler()
-consoleHandler.setFormatter(logFormatter)
-consoleHandler.setLevel(logging.INFO)
-logging.getLogger().addHandler(consoleHandler)
-import warnings
-logging.getLogger('sherpa.plot').setLevel(logging.ERROR)
-warnings.filterwarnings("ignore", message='displayed errorbars')
-
-ids = []
-
-id = 1
-filename = os.environ['FILENAME']
-#filename = 'combined_src.pi'
-prefix = filename + '_out_'
-load_pha(id, filename)
-ignore_id(id, None, None)
-notice_id(id,float(os.environ.get('ELO', 0.5)),float(os.environ.get('EHI',8)))
-ids.append(id)
-
-set_analysis(id, 'ener', 'counts')
-set_xlog()
-set_ylog()
-set_stat('cstat')
-set_xsabund('wilm')
-set_xsxsect('vern')
-
-
-id = ids[0]
-galabso = bxa.auto_galactic_absorption(id)
-galabso.nH.freeze()
-
-# Models available at https://doi.org/10.5281/zenodo.602282
-load_table_model("torus", '/home/user/Downloads/specmodels/uxclumpy-cutoff.fits')
-load_table_model("scat", '/home/user/Downloads/specmodels/uxclumpy-cutoff-omni.fits')
-# the limits correspond to fluxes between Sco X-1 and CDFS7Ms faintest fluxes
-srclevel = Parameter('src', 'level', 0, -8, 3, -20, 20) 
-print('combining components')
-model = torus + scat
-print('linking parameters')
-torus.norm = 10**srclevel
-srcnh = Parameter('src', 'nH', 22, 20, 26, 20, 26)
-torus.nh = 10**(srcnh - 22)
-scat.nh = torus.nh
-scat.nh = torus.nh
-print('setting redshift')
-redshift = Parameter('src', 'z', 1, 0, 10, 0, 10) 
-torus.redshift = redshift
-scat.redshift = redshift
-scat.phoindex = torus.phoindex
-
-scat.ecut = torus.ecut
-scat.theta_inc = torus.theta_inc
-scat.torsigma = torus.torsigma
-scat.ctkcover = torus.ctkcover
-softscatnorm = Parameter('src', 'softscatnorm', -2, -7, -1, -7, -1)
-scat.norm = 10**(srclevel + softscatnorm)
-
-
-
-
[…]
