### 2021-03-30 6df21b8
improve description of previous works and differences
