+++ paper.bib
@@ -199,3 +199,18 @@
   review = {Bayesian X-ray spectral analysis using MCMC},
   timestamp = {2012.08.26}
 }
+@ARTICLE{GoodmanWeare,
+  author = {{Goodman}, J. and {Weare}, J.},
+  title = {{Ensemble samplers with affine invariance}},
+  journal = {Communications in Applied Mathematics and Computational Science,
+	Vol.~5, No.~1, p.~65-80, 2010},
+  year = {2010},
+  volume = {5},
+  pages = {65-80},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+  adsurl = {http://adsabs.harvard.edu/abs/2010CAMCS...5...65G},
+  doi = {10.2140/camcos.2010.5.65},
+  keywords = {Markov chain Monte Carlo, affine invariance, ensemble samplers },
+  owner = {user},
+  timestamp = {2017.09.12}
+}
