+++ paper.md
@@ -12,13 +12,13 @@ authors:
     orcid: 0000-0003-0872-7098
     affiliation: "1, 2, 3, 4"
 affiliations:
- - name: Max Planck Institute for Extraterrestrial Physics, Giessenbachstrasse, 85741 Garching, Germany. 
+ - name: Max Planck Institute for Extraterrestrial Physics, Giessenbachstrasse, 85741 Garching, Germany.
    index: 1
- - name: Millenium Institute of Astrophysics, Vicuña MacKenna 4860, 7820436 Macul, Santiago, Chile . . . 
+ - name: Millenium Institute of Astrophysics, Vicuña MacKenna 4860, 7820436 Macul, Santiago, Chile . . .
    index: 2
- - name: Pontificia Universidad Católica de Chile, Instituto de Astrofísica, Casilla 306, Santiago 22, Chile. 
+ - name: Pontificia Universidad Católica de Chile, Instituto de Astrofísica, Casilla 306, Santiago 22, Chile.
    index: 3
- - name: Excellence Cluster Universe, Boltzmannstr. 2, D-85748, Garching, Germany  
+ - name: Excellence Cluster Universe, Boltzmannstr. 2, D-85748, Garching, Germany
    index: 4
 
 date: 22 January 2021
@@ -28,32 +28,38 @@ bibliography: paper.bib
 
 # Summary
 
-BXA connects the X-ray specral analysis environments Xspec 
+BXA connects the X-ray spectral analysis environments Xspec
 [@XSPEC] and Sherpa [@Freeman2001]
 with modern Bayesian inference algorithms based on nested sampling.
 This allows estimation of parameter probability distributions and
-model comparison with Bayes factors. Because BXA is a plugin,
-users can build and fit the full range of models 
-available in Xspec and Sherpa. 
-BXA also comes with convenience functionality for backgrounds 
+model comparison with Bayes factors. Because BXA is a plug-in,
+users can build and fit the full range of models
+available in Xspec and Sherpa.
+BXA also comes with convenience functionality for backgrounds
 of many X-ray missions, which allow extracting more information from low-count data.
 
 # Statement of need
 
 In X-ray spectroscopy, the instrument responses are often not trivial,
 which means that forward folding is required to test models against data.
-Judging realistic physical models, which can be degenerate, 
+Judging realistic physical models, which can be degenerate,
 against data, which sometimes consist of few photon counts, is a challenge.
 Bayesian inference provides a good solution in this setting.
 
-For parameter estimation, Bayesian inference updates an initial state of information 
+For parameter estimation, Bayesian inference updates an initial state of information
 using the data and a model into a new state (posterior probability distribution).
-In practice, Bayesian inference algorithms perform an approximation using sampling.
-Before BXA, Bayesian X-ray spectral analyses have been limited to 
-versions of Markov Chain Monte Carlo (MCMC) [e.g., @Dyk2001]
-that were difficult to initialise and tune.
-MCMC also has limitations when models have multiple parameter peaks, which is
+In practice, Bayesian inference algorithms approximate the posterior by Monte Carlo sampling.
+Before BXA, Bayesian X-ray spectral analyses have been implemented with
+Markov Chain Monte Carlo (MCMC).
+The variants include Metropolis random walks [e.g., @Dyk2001] and
+affine-invariant ensembles [@GoodmanWeare].
+These are difficult to initialise and/or tune, and require complicated
+checks of the chain convergence.
+MCMC tends to be computationally costly, and difficult to parallelise.
+MCMC has limitations when models have multiple posterior peaks, which is
 common in the additive models typically considered in X-ray astronomy.
+For these reasons, Bayesian X-ray analyses, as implemented in existing
[…]
