### 2021-03-30 abe6a19
typo: docs folder is called doc, make sure it is cleaned
