### 2021-03-30 404ea5f
also point in sherpa to a custom prior function
