+++ doc/xspec-analysis.rst
@@ -37,7 +37,7 @@ unit variables unto the values needed for each parameter.
 See the `UltraNest documentation on priors <https://johannesbuchner.github.io/UltraNest/priors.html>`_ 
 for more details about this concept.
 The script `examples/xspec/example_advanced_priors.py <https://github.com/JohannesBuchner/BXA/blob/master/examples/xspec/example_advanced_priors.py>`_ 
-gives an example of such a custom prior function.
+gives an example of such a custom prior function (`my_custom_prior`).
 
 API information:
 
