### 2021-03-30 JohannesBuchner/BXA#24 @JohannesBuchner
> The documentation refers to "prior transformations" throughout the docs. This is a reference to transforming the nested sampling focus on transforming all priors from a unit cube.
>
> This could be confusing as users who do not realize that this is specifically for nested sampling and not a form of MCMC that is provided in both XSPEC (https://asd.gsfc.nasa.gov/XSPECwiki/Markov_Chain_Monte_Carlo) and Sherpa (https://sherpa.readthedocs.io/en/latest/mcmc/index.html). It might be useful at the introduction (since it is not in the name) to clarify this? The connection to nested sampling is clearly there, but a note or something similar could be more instructive.

Thank you for pointing out this confusion.
The text was edited in the paper to better introduce the existing capabilities of xspec and sherpa.
In the documentation, the transforms are now better described on the xspec https://johannesbuchner.github.io/BXA/xspec-analysis.html#defining-priors  and sherpa pages https://johannesbuchner.github.io/BXA/sherpa-analysis.html#defining-priors . Links to the UltraNest docs which give a more thorough introduction were added.

> Perhaps is would useful to define what a transformation is here:
> https://github.com/JohannesBuchner/BXA/blob/master/bxa/xspec/solver.py#L104

Thanks, more information was added there.

> **There are references to pymultinest throughout the docs.**

These were all replaced now, thank you!

> What are all the available priors and can a user add custom priors?

This has been clarified in the prior documentation sections on the xspec & sherpa pages. The priors are quite basic, but custom prior functions can be used.

> In the examples, a Gaussian prior must be specified with limits. Is it not possible to have an unbounded Gaussian prior?

It is possible to set the limits very wide, but BXA cannot set the values to outside the parameter limits defined in xspec/sherpa.

> It seems there is a wealth of functionality with priors in the examples that is perhaps not documented and a naive user could miss these capabilities.

This is improved now by providing links to these advanced examples.

Thanks!
