### 2021-03-30 JohannesBuchner/BXA#25 @JohannesBuchner
Thank you for catching this. The paper and doc are now updated to state:

> BXA shines especially
>
>     when systematically analysing a large data-set, or
>
>     when comparing multiple models
>
>     when analysing low counts data-set with realistic models
>
> because its robust and unsupervised fitting algorithm explores even complicated parameter spaces in an automated fashion. Unlike existing approaches, the user does not need to apply problem-specific algorithm parameter tuning, initialise to good starting points and check for convergence, allowing building automated analysis pipelines.

The main point is that automatic analyses can be performed, which is difficult with the available MCMC implementations (because they need to be combined with custom initialisation and termination criteria).

While a numerical study that compares the performance of MCMC flavors+init+convergence checks and BXA on a example problem is beyond the scope of this paper/documentation, the paper text was updated to go into more detail on the differences to existing MCMC implementations. I focus on highlighting _the difference in nature_ of the approaches (complete solution with initialisation and termination, no algorithm parameters need to be tuned to the problem). It is certainly possible to do low-count analyses also with the existing MCMC capabilities, at least with mono-modal gaussian-like posteriors.
