### 2021-04-06 #3045 @grburgess
I think many of the issues were addressed, but some of them are still open. I have not had a chance to reinstall everything.

 I still think it would be nice to have a badge in the repo detailing which operating systems BXA does work on as I was never able to get its major dependencies installed on Mac OS. This could be handles through CI badging and testing. This would help to differentiate the software as be "feature complete" and not

>designed for maintainable extension (not one-off modifications of existing tools). “Minor utility” packages, including “thin” API clients, and single-function packages are not acceptable.

as stated in the review checklist. Currently, BXA adds a different method of obtaining posterior samples to two x-ray analysis packages which already provide a similar functionality as well as some enhanced plotting. But it completely relies on those packages being available for install on a given platform. I'm not sure what the editorial (@jgostick ) opinion is on this, so I'll just leave it linked here: (https://github.com/JohannesBuchner/BXA/issues/22) It is also related to the CI and testing issue but I think these both relate to the same point.

Also, a few of the claims in the paper should be cited or checked regarding the comparison to the included posterior estimation in the core packages. The nested sampling algorithm that BXA attaches to XSPEC and Sherpa, also has tuning parameters (e.g. dlogz, frac_remain etc.) as well as initialization tuning (e.g. warm starts . This will help those reading the paper to understand that these are different methods and one is not actually an improvement on the the other unless there are citations that can show that, for example, MCMC does not work in these situations, though this is not my own personal experience in general. I believe this exists for the affine-invariant flavors of MCMC, but it is not a general statement about MCMC. Moreover, as there do not exist analytic convergence criteria for nested sampling (as there for for Markov transitions), the statement in the paper can be misleading as there exists no posterior  convergence checks for nested sampling other than a stoping criteria on the accuracy of the marginal likelihood. Thus, I think it can be pointed out that BXA attaches a nested sampling algorithm to XSPEC and Sherpa to provide the ability to use that algorithm in addition to what is in available in those packages and point out that nested sampling can estimate the marginal likelihood and explore multimodal posteriors without claims about convergence checks and tuning parameters.

Other than those two points, everything else is satisfied.
