### 2021-04-06 de94365
add information about supported operating systems
