### 2021-04-06 #3045 @cescalara
Hi @JohannesBuchner, I had another look and some final comments:
* Everything works fine on linux (ubuntu) but I haven't managed to install the necessary dependencies (Xspec/sherpa) on macOS either following what instructions I could find online. It seems like it may be possible for certain versions, but is not so simple, and definitely much harder than my experience getting set up on linux. I think this is fine but should be flagged somewhere like @grburgess suggested so that people know they will have to do some extra effort/googling if they want to get going on Mac from scratch.
* I am no nested sampling expert, but I imagine some of the algorithm parameters could be tuned to optimise things for a specific application, even if you have picked robust defaults. When working with MCMC packages, I only usually think about tuning parameters should any issues arise in the sampling with default inputs, so the experience isn't massively different? Additionally, I'm sure things can occasionally go wrong (e.g. the discussion in https://arxiv.org/abs/1804.06406) so it can be good to encourage users to check out the trace plots and have a think about what is going on rather than encourage "black box" usage. But perhaps, I have misinterpreted what you are saying in the paper.
