### 2021-04-06 JohannesBuchner/BXA#22 @JohannesBuchner
You mentioned here the MacOS install attempt:

> After completing the install for CIAO with the following steps:
>
> bash ciao_install
> pip install sherpa # (installs numpy)
> python setup.py install # inside BXA folder cloned from GitHub

Please try again in a clean environment with:

```
bash ciao_install
source /path/to/soft/ciao-4.13/bin/ciao.bash
python3 setup.py install # inside BXA folder cloned from GitHub
```
(This is following https://cxc.cfa.harvard.edu/ciao/download/ciao_install.html)
