### 2021-04-06 9d27743
tone down claims on superiority over MCMC
