+++ paper.bib
@@ -214,3 +214,25 @@
   owner = {user},
   timestamp = {2017.09.12}
 }
+@ARTICLE{Speagle2020,
+  author = {{Speagle}, Joshua S.},
+  title = {{DYNESTY: a dynamic nested sampling package for estimating Bayesian
+	posteriors and evidences}},
+  journal = {\mnras},
+  year = {2020},
+  volume = {493},
+  pages = {3132-3158},
+  number = {3},
+  month = apr,
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+  adsurl = {https://ui.adsabs.harvard.edu/abs/2020MNRAS.493.3132S},
+  archiveprefix = {arXiv},
+  doi = {10.1093/mnras/staa278},
+  eprint = {1904.02180},
+  file = {arXiv v1:Speagle2020-eprintv1.pdf:PDF},
+  keywords = {methods: data analysis, methods: statistical, Astrophysics - Instrumentation
+	and Methods for Astrophysics, Statistics - Computation},
+  owner = {user},
+  primaryclass = {astro-ph.IM},
+  timestamp = {2020.06.17}
+}
