### 2021-04-06 bcadf9f
doc: tone down black box usage claims
