### 2021-04-06 e4ff24c
remove MCMC parallelisation claim
