### 2021-04-06 84972ab
add more information on diagnosing runs
