+++ paper.bib
@@ -236,3 +236,39 @@
   primaryclass = {astro-ph.IM},
   timestamp = {2020.06.17}
 }
+@ARTICLE{Higson2019,
+  author = {{Higson}, Edward and {Handley}, Will and {Hobson}, Michael and {Lasenby},
+	Anthony},
+  title = {{NESTCHECK: diagnostic tests for nested sampling calculations}},
+  journal = {\mnras},
+  year = {2019},
+  volume = {483},
+  pages = {2044-2056},
+  number = {2},
+  month = {Feb},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+  adsurl = {https://ui.adsabs.harvard.edu/abs/2019MNRAS.483.2044H},
+  archiveprefix = {arXiv},
+  doi = {10.1093/mnras/sty3090},
+  eprint = {1804.06406},
+  file = {arXiv v1:Higson2019-eprintv1.pdf:PDF},
+  keywords = {methods: data analysis, methods: numerical, methods: statistical,
+	Statistics - Computation, Astrophysics - Cosmology and Nongalactic
+	Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics,
+	Physics - Data Analysis, Statistics and Probability},
+  owner = {user},
+  primaryclass = {stat.CO},
+  review = {contributions:
+	
+	
+	* diagram - show uncertainty in posterior using bootstrapping
+	
+	* diagram - logX vs samples
+	
+	* multirun - BS vs multirun mean uncertainties per parameter
+	
+	* pairtest - KS test of means from two runs
+	
+	* diagnostic -},
+  timestamp = {2019.10.05}
+}
