### 2021-04-06 #3045 @JohannesBuchner
Thank you for the useful comments. I address the two remaining issues below.

Regarding which operating systems BXA is supported on:

A section was added to the installation instructions at https://github.com/JohannesBuchner/BXA#installation. It refers the user to supported operating systems.

The tested operating systems of the CI is also mentioned there now (currently only Ubuntu Linux).

I am sorry that the installation of ciao and heasoft on MacOS proves difficult.

Regarding the differentiation of MCMC and NS:

> Also, a few of the claims in the paper should be cited or checked regarding the comparison to the included posterior estimation in the core packages. The nested sampling algorithm that BXA attaches to XSPEC and Sherpa, also has tuning parameters (e.g. dlogz, frac_remain etc.) as well as initialization tuning (e.g. warm starts . This will help those reading the paper to understand that these are different methods and one is not actually an improvement on the the other unless there are citations that can show that, for example, MCMC does not work in these situations, though this is not my own personal experience in general. I believe this exists for the affine-invariant flavors of MCMC, but it is not a general statement about MCMC.

> I am no nested sampling expert, but I imagine some of the algorithm parameters could be tuned to optimise things for a specific application, even if you have picked robust defaults. When working with MCMC packages, I only usually think about tuning parameters should any issues arise in the sampling with default inputs, so the experience isn't massively different?

The text was modified to remove the claims on tuning parameters (the relevant ones for NS would be the number of live points and parameters of the proposal distribution, which in the case of the MLFriends algorithm used here, is the number of bootstrapping rounds, as well as the termination criterion).
Instead, the "statement of need" first present MCMC, and then nested sampling as an algorithm with different properties, without making strong/blanket superiority claims. An example reference to support the speed comparison was added.

The initialization is indeed different (MCMC requiring a starting point, NS starting the exploration from prior sampling), so this was kept, as it is a relevant robustness benefit of NS in X-ray spectroscopy. Warm start is an optional UltraNest feature never used by BXA, so it is not relevant to mention.

The text was also modified to be clearer that the statements only refer to the available MCMC implementations, not to all possible MCMC algorithms (for example, HMC with parallel tempering may have different properties).

> Moreover, as there do not exist analytic convergence criteria for nested sampling (as there for for Markov transitions), the statement in the paper can be misleading as there exists no posterior convergence checks for nested sampling other than a stoping criteria on the accuracy of the marginal likelihood. Thus, I think it can be pointed out that BXA attaches a nested sampling algorithm to XSPEC and Sherpa to provide the ability to use that algorithm in addition to what is in available in those packages and point out that nested sampling can estimate the marginal likelihood and explore multimodal posteriors without claims about convergence checks and tuning parameters.

> I am no nested sampling expert, but I imagine some of the algorithm parameters could be tuned to optimise things for a specific application, even if you have picked robust defaults. When working with MCMC packages, I only usually think about tuning parameters should any issues arise in the sampling with default inputs, so the experience isn't massively different? Additionally, I'm sure things can occasionally go wrong (e.g. the discussion in https://arxiv.org/abs/1804.06406) so it can be good to encourage users to check out the trace plots and have a think about what is going on rather than encourage "black box" usage. But perhaps, I have misinterpreted what you are saying in the paper.

I agree, and discussion of tuning parameters were removed. However, when the computation is stopped (termination) is importantly different in MCMC and NS. In MCMC, one would typically assess the convergence of the chain(s) and decide whether they have been run long enough, and if not, either resume the chains or run anew for much longer. In NS, the algorithm itself converges, in the sense that after some iterations any newly sampled points do not contribute anything to the posterior (as the volume becomes negligible. This is what was meant here.

The other meaning of convergence -- convergence of the posterior sample distribution to the actual posterior distribution -- is also important. Typically in MCMC it is assessed by comparing multiple chains and their sections (e.g., "r-hat").
The nestcheck paper is an important contribution in this direction for NS. A section was added to the paper to advise users of diagnosing poor runs, based on excess variance in the marginal posterior and ln(Z) value.
The algorithm implemented by UltraNest is designed to diagnose itself and adjust automatically in a very conservative fashion (https://arxiv.org/abs/1407.5459). Further diagnostics (e.g., https://arxiv.org/abs/2006.03371) are available, but are a topic of current research, and more research and practical experience is needed to recommend them.

The text was modified to remove language encouraging "black box" usage.

The text changes can be seen in this diff:
https://github.com/JohannesBuchner/BXA/compare/6df21b8..joss-paper#diff-0403ef06adf405f7b310b4518bd6a3559854f54c61676f676ce9cbfee7172ab6

I appreciate the comments, and please let me know if you have further suggestions.
