### 2021-04-08 03710ed
fix bug in sherpa PCA implementation & solver .fit access
