### 2021-04-08 edfb1ef
[ci] make ci more stringent
