### 2021-04-08 ee54a42
[ci] deleting rebinned model should not cause failure
