+++ examples/sherpa/runall.sh
@@ -6,8 +6,10 @@ coverage run -p example_simplest.py || exit 1
 coverage run -p example_pcabackground.py || exit 2
 coverage run -p example_automatic_background_model.py || exit 3
 
-rm testmodel.npz  && coverage run -p example_rebinning.py || exit 4
-coverage run -p example_rebinning.py || exit 5
+rm -f testmodel.npz
+coverage run -p example_rebinning.py || exit 4
+ls testmodel.npz || exit 5
+coverage run -p example_rebinning.py || exit 6
 
 coverage run -p model_compare.py superfit/ tbabs_noz/
 
