### 2021-04-08 8c1b558
[ci] make running runall.sh more verbose
