### 2021-04-09 3559565
[ci] missed tqdm as requirement
