### 2021-04-09 9512a14
[ci] allow convert command to read pdfs
