+++ .travis.yml
@@ -32,6 +32,8 @@ before_cache:
 install:
   - sudo apt-get install -y python3-numpy python-numpy python3-dev python-tk python3-pip python3-setuptools
   - sudo pip3 install --upgrade pip==20.3.4
+  # allow convert to read PDF files:
+  - sudo sed -i '/PDF/s/none/read|write/' /etc/ImageMagick-6/policy.xml
   - which python || true
   - which python3 || true
   - export MYCACHE=$HOME/Downloads/$TOOL
