### 2021-04-09 742b2a6
[ci] install ghostscript for convert
