### 2021-04-26 f999a8e
set xagnfitter inclination
