### 2021-04-26 #3045 @JohannesBuchner
> Here is a list of what to do next:
>
>     * [ ]  Double check authors and affiliations (including ORCIDs)

Thanks!

>     * [ ]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

v4.0.5  at  https://github.com/JohannesBuchner/BXA/releases/tag/v4.0.5

>     * [ ]  Archive the release on Zenodo/figshare/etc and post the DOI here.

Release:
https://zenodo.org/record/4721575

DOI:
10.5281/zenodo.4721575

>     * [ ]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Done.
