### 2021-05-04 #3045 @jgostick
I wasn't paying enough attention.  The dialog between the author and @grburgess was so thorough that I just subconsciously checked the boxes in my head.
