### 2021-05-05 #3045 @jgostick
I'm really sorry about this @kthyng...I will pay more attention in future.  I actually ask Arfon to give whedon a command like 'generate editor checklist'...he seemed to think it was a good idea so we'll see...
