### 2021-05-05 #3045 @kthyng
@jgostick No worries! I would like there to be an auto check for the check list myself.

Thanks all! I think all the nooks and crannies are covered now.

Congrats on your new publication @JohannesBuchner and thanks to editor @jgostick and reviewers @cescalara and @grburgess for your hard work and expertise!!
